Added amResources and ImageLoader class
This will make the game more efficient and it's easier to access every
image, music or whatever we need for the game.
The ImageLoader class reads
images so you don't need to duplicate code.

Instead of loading a image everytime the game loop calls run(). The
game loads every image one time and can be accessed through
Recourses."image name";