Update WorldBlocks.java
Added the WorldBlocks class which handles and stores every block in the game. The blocks are also given a unique id which is used when creating the world in the text file. For example if a dirt block has the id 0, writing a 0 in the text file would render the dirt block